feat: superadmin dashboard, user management, bug fixes & CSS design system
- Fix migration PKs: id_pengguna, id_cabang (was generic `id`)
- Fix FK types & unique validation to use correct column names
- Add migration for missing tables (kursus, materi, tugas, kuis, dll)
- Add CabangSeeder before UserSeeder in DatabaseSeeder
- Fix LoginLog relationship FK (id_pengguna)
- Record login/logout to login_logs in both AuthControllers
- Fix SuperAdmin controllers: all references to id -> id_pengguna/id_cabang
- Fix frontend: setBranches/setUsers response unwrapping (r.data.data)
- Add CSS design system variables (bg-card, text-primary, border-theme, etc)
- Login page always renders in light mode, restores dark pref on unmount

+++ backend/database/migrations/2026_03_09_023921_create_login_logs_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('login_logs', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id_pengguna')->on('pengguna')->onDelete('cascade');
             $table->string('ip_address')->nullable();
             $table->timestamp('logged_in_at')->useCurrent();
